backend/requirements-training.txt
# Python dependencies for model training and inference
torch==2.1.0
transformers==4.35.0
pymongo==4.5.0
numpy==1.24.0
pandas==2.1.0
datasets==2.14.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
